requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.4
pycryptodome~=3.23.0
flask~=3.1.0
flask-cors~=5.0.1